First draft of damage calculation
Ok, i've tried doing it like this, should work. Couple of things might
need changing. Since I added elemental type as it's own variable, the
specific damage % needs to be gotten at easily based onthe value of the
elemental variable.  ie.  1 for cold, 2 for fire, etc. This will
probably need a small change in the way we get the input, maybe by using
an array to loop through to on the input side, or something.
Also, basically the calculationcode for the 3 active skills is exactly
the same, this might be simplified some more.
For number of targets, is there anyway to cap it at say 20? It's
pointless beyond 11 already, I just need to put in some number in the
Rune object, and this seems the easiest way.