Changed breakpoint calc dropdown and button styling
I didn't like it with the dropdown being so wide, and the button so
small if the the text in the dropdown isn't centered. I couldn't figure
out how to center it, maybe that would be better still, so I styled the
same way as fields above it. I just blanked out the original code, so
you decide what to do.